chore: Add ImageGeneratorTool to the brain integration module
