+++ backend/modules/tools/__init__.py
@@ -1,0 +1,2 @@
+from .image_generator import ImageGeneratorTool
+
